+++ src/Core/Query/QueryParameters.php
@@ -23,6 +23,7 @@ use Illuminate\Contracts\Support\Arrayable;
 use Illuminate\Http\Request;
 use Illuminate\Support\Enumerable;
 use LaravelJsonApi\Contracts\Query\QueryParameters as QueryParametersContract;
+use LaravelJsonApi\Contracts\Schema\Schema;
 use LaravelJsonApi\Core\Support\Arr;
 use UnexpectedValueException;
 use function array_key_exists;
@@ -52,9 +53,9 @@ class QueryParameters implements QueryParametersContract, Arrayable
     private ?array $pagination;
 
     /**
-     * @var array|null
+     * @var FilterParameters|null
      */
-    private ?array $filters;
+    private ?FilterParameters $filters;
 
     /**
      * @var array
@@ -136,7 +137,7 @@ class QueryParameters implements QueryParametersContract, Arrayable
             array_key_exists('fields', $value) ? FieldSets::cast($value['fields']) : null,
             array_key_exists('sort', $value) ? SortFields::cast($value['sort']) : null,
             array_key_exists('page', $value) ? $value['page'] : null,
-            array_key_exists('filter', $value) ? $value['filter'] : null,
+            array_key_exists('filter', $value) ? FilterParameters::cast($value['filter']) : null,
             $unrecognised,
         );
     }
@@ -161,7 +162,7 @@ class QueryParameters implements QueryParametersContract, Arrayable
      * @param FieldSets|null $fieldSets
      * @param SortFields|null $sortFields
      * @param array|null $page
-     * @param array|null $filters
+     * @param FilterParameters|null $filters
      * @param array|null $unrecognised
      */
     public function __construct(
@@ -169,7 +170,7 @@ class QueryParameters implements QueryParametersContract, Arrayable
         FieldSets $fieldSets = null,
         SortFields $sortFields = null,
         array $page = null,
-        array $filters = null,
+        FilterParameters $filters = null,
         array $unrecognised = null
     ) {
         $this->includePaths = $includePaths;
@@ -197,7 +198,7 @@ class QueryParameters implements QueryParametersContract, Arrayable
     }
 
     /**
-     * @return IncludePaths|null
+     * @inheritDoc
      */
     public function includePaths(): ?IncludePaths
     {
@@ -226,7 +227,7 @@ class QueryParameters implements QueryParametersContract, Arrayable
     }
 
     /**
-     * @return FieldSets|null
+     * @inheritDoc
      */
     public function sparseFieldSets(): ?FieldSets
     {
@@ -289,7 +290,7 @@ class QueryParameters implements QueryParametersContract, Arrayable
     }
 
     /**
-     * @return SortFields|null
+     * @inheritDoc
      */
     public function sortFields(): ?SortFields
     {
@@ -318,7 +319,7 @@ class QueryParameters implements QueryParametersContract, Arrayable
     }
 
     /**
-     * @return array|null
+     * @inheritDoc
      */
     public function page(): ?array
     {
@@ -351,9 +352,9 @@ class QueryParameters implements QueryParametersContract, Arrayable
     }
 
     /**
-     * @return array|null
+     * @inheritDoc
      */
-    public function filter(): ?array
+    public function filter(): ?FilterParameters
     {
         return $this->filters;
     }
@@ -361,12 +362,12 @@ class QueryParameters implements QueryParametersContract, Arrayable
     /**
      * Set filters.
      *
-     * @param Arrayable|array|null $filters
+     * @param mixed|null $filters
      * @return $this
      */
     public function setFilters($filters): self
     {
-        $this->filters = is_null($filters) ? null : collect($filters)->toArray();
+        $this->filters = FilterParameters::nullable($filters);
 
         return $this;
     }
@@ -427,8 +428,8 @@ class QueryParameters implements QueryParametersContract, Arrayable
             $query['fields'] = $this->fieldSets->toArray();
         }
 
-        if (is_array($this->filters) && !empty($this->filters)) {
-            $query['filter'] = $this->filters;
+        if ($this->filters && $this->filters->isNotEmpty()) {
+            $query['filter'] = $this->filters->toArray();
         }
 
         if ($this->includePaths && $this->includePaths->isNotEmpty()) {
@@ -459,8 +460,8 @@ class QueryParameters implements QueryParametersContract, Arrayable
             $query['fields'] = $this->fieldSets->toArray();
         }
 
-        if (is_array($this->filters) && !empty($this->filters)) {
-            $query['filter'] = $this->filters;
+        if ($this->filters && $this->filters->isNotEmpty()) {
+            $query['filter'] = $this->filters->toArray();
         }
 
         if ($this->includePaths && $this->includePaths->isNotEmpty()) {
@@ -478,5 +479,34 @@ class QueryParameters implements QueryParametersContract, Arrayable
         ksort($query);
 
         return $query;
+    }
+
+    /**
+     * Get a new query parameter object with valid values for the provided schema.
+     *
+     * @param Schema $schema
+     * @return $this
+     */
+    public function forSchema(Schema $schema): self
+    {
+        $includePaths = IncludePaths::cast($this->includePaths)->forSchema($schema);
+        $sortFields = SortFields::cast($this->sort)->forSchema($schema);
+        $filters = FilterParameters::cast($this->filters)->forSchema($schema);
+        $pagination = null;
+
+        if ($this->pagination && $paginator = $schema->pagination()) {
+            $pagination = collect($this->pagination)
+                ->only($paginator->keys())
+                ->all();
+        }
+
+        return new static(
+            $includePaths->isNotEmpty() ? $includePaths : null,
+            $this->fieldSets,
+            $sortFields->isNotEmpty() ? $sortFields : null,
+            !empty($pagination) ? $pagination : null,
+            $filters->isNotEmpty() ? $filters : null,
+            $this->unrecognised,
+        );
     }
 }